Fix duplicate code, POS_CHECKSUM overflow, and button state references

+++ include/obstacle_config.h
@@ -43,7 +43,7 @@ namespace ObstacleConfig {
     constexpr uint8_t POS_LENGTH = 5;               // Length (2 bytes, little-endian)
     constexpr uint8_t POS_SYSTEM_TIME = 7;          // System time (4 bytes, little-endian)
     constexpr uint8_t POS_MATRIX_DATA = 11;         // Matrix data start (384 bytes)
-    constexpr uint8_t POS_CHECKSUM = 395;           // Checksum byte
+    constexpr uint16_t POS_CHECKSUM = 395;          // Checksum byte (uint16_t for value >255)
     constexpr uint16_t POS_RESERVED = 396;          // Reserved (4 bytes)
     
     // Distance thresholds (mm)
